# backend/internal/swaggerdocs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "NextChapter is a self-hosted progress tracker for manhwa, manhua, and web novels.\nOpen registration — anyone can create an account via /auth/register.\n\nAuth: session cookie (\"nc_session\") OR Authorization: Bearer \u003capi token\u003e.",
        "title": "NextChapter API",
        "contact": {
            "name": "NextChapter"
        },
        "license": {
            "name": "See LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates the supplied credentials. On success the server sets the nc_session cookie and returns the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in with username and password",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the session row (if cookie-authed) and clears the nc_session cookie. Idempotent.",
                "tags": [
                    "auth"
                ],
                "summary": "Log out the current session",
                "responses": {
                    "204": {
                        "description": "no content"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Return the currently authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Open registration: anyone can create an account. On success the server sets the nc_session cookie and returns the new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user account",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "registration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Registration"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/tokens": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the plaintext token exactly once on the `token` field. The server stores only the hash; subsequent reads omit `token`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Mint a new API token for the current user",
                "parameters": [
                    {
                        "description": "token metadata",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewToken"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.APIToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/tokens/{id}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revoke an API token",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/entries": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entries"
                ],
                "summary": "List per-site reading positions",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "filter by series id",
                        "name": "series_id",
                        "in": "query"
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "page offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EntryList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/entries/capture": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upserts on (user, site_host, series_slug). The handler lowercases site_host and strips a leading `www.` before persisting. If no entry exists for that key, the caller must supply either `series_id` (existing) or `new_series_title` (the handler will materialise a series row); otherwise 422. Returns 201 on first capture for the (host,slug) key, 200 when an existing entry was advanced.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entries"
                ],
                "summary": "Capture or advance a reading position",
                "parameters": [
                    {
                        "description": "captured reading position",
                        "name": "capture",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EntryCapture"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "advanced an existing entry",
                        "schema": {
                            "$ref": "#/definitions/models.Entry"
                        }
                    },
                    "201": {
                        "description": "created a new entry",
                        "schema": {
                            "$ref": "#/definitions/models.Entry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/entries/{id}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "entries"
                ],
                "summary": "Forget a single per-site reading position",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "entry id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pointer fields use the absent/present binary: missing fields are left unchanged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entries"
                ],
                "summary": "Adjust an existing entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "entry id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "fields to update",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EntryPatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Entry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/series": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "List the user's tracked series",
                "parameters": [
                    {
                        "enum": [
                            "reading",
                            "completed",
                            "on_hold",
                            "dropped",
                            "plan_to_read"
                        ],
                        "type": "string",
                        "description": "filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "page offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "AND-filter; repeatable, lowercase tag",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SeriesList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Track a new series",
                "parameters": [
                    {
                        "description": "series metadata",
                        "name": "series",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SeriesNew"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Series"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/sites": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns two parallel lists: `rules` is the per-user site_rule rows (seeded from compiled-in defaults at registration); `tracked_hosts` is the distinct site_host values across the caller's entries. The two are independent — a host can be tracked without a rule, and a rule can exist without any captures.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "List the user's site rules plus the hosts they've captured chapters on",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SiteList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.ErrorDetail"
                }
            }
        },
...
        "models.APIToken": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "last_used_at": {
                    "type": "string"
                },
                "token": {
                    "description": "Raw is the plaintext token, surfaced exactly once on\nPOST /auth/tokens. The server stores only the hash; subsequent\nreads return APIToken values with Raw == \"\" (omitempty drops the\nkey entirely from the wire payload).",
                    "type": "string"
                }
            }
        },
        "models.Credentials": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.NewToken": {
            "type": "object",
            "required": [
                "label"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "label": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                }
            }
        },
        "models.Registration": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 256,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "nc_session",
            "in": "cookie"
        }
    }
}